fix: add vertexes to avoid dunders
Since we create a cyclic graph, the vertexes should be immutable objects. In
other words, they should not implement the __hash__ function. A better solution
would be to store edges and nodes separately by using networkx, which was not
taken into account for this commit.

Added to this, defining __eq__ in terms of __hash__, that's a bug
waiting to happen.

This commit takes into account the initial solution and assumes that a
vertex is hashable, even if it should not be case.

Thus, from the moment an object is hashable, we should use @attr.s to
set up all the good dunders (e.g., __eq__, __hash__) for us.

NOTE: if we need to change the name of a vertex, we could still use
`vertex = attrs.evolve(v, name="test")`

The cool thing with frozen is that we can use the whole object as a key
in dicts and things. However, we should discuss if vertexes need a name
and an id.